refactor: eliminate remaining duplications from export pipeline
Extract _load_bubble_map/_load_project_layouts_map/_load_code_block_diff_map
to services/workspace_db.py; add cursor_ide_chat_to_markdown to
utils/cursor_md_exporter.py; remove private _slug() copies from
api/export_api.py and utils/cursor_md_exporter.py.
